display basic SRM color palette and let user select one for recipe

+++ BeerRecipeCore/src/main/kotlin/com/jlafshari/beerrecipecore/Recipe.kt
@@ -1,4 +1,4 @@
 package com.jlafshari.beerrecipecore
 
 data class Recipe(val id: String, val style: Style, val size: Double, val name: String, val abv: Double,
-                  val colorSrm: Double)
+                  val colorSrm: Int)
